+++ src/pymetropolis/metro_results/demand/__init__.py
@@ -1,5 +1,5 @@
-from .files import TripResultsFile
-from .postprocess import TripResultsStep
+from .files import RouteResultsFile, TripResultsFile
+from .postprocess import RouteResultsStep, TripResultsStep
 
-DEMAND_RESULTS_FILES = [TripResultsFile]
-DEMAND_RESULTS_STEPS = [TripResultsStep]
+DEMAND_RESULTS_FILES = [TripResultsFile, RouteResultsFile]
+DEMAND_RESULTS_STEPS = [TripResultsStep, RouteResultsStep]